app/requirements.txt
fastapi
uvicorn[standard]
google-cloud-pubsub
python-multipart